requirements.txt
requests
python-dotenv
PyYAML
pika
aiosmtpd
loguru
pydantic